requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Pygments==2.19.1
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.1
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.1
wheel==0.46.0
